added axios and bootstrap via npm, fixed responsivness of app, started on the event listener
So far, I have been able to get my event listener to hear my submit, and it alerts me that it's retrieving the weather. Next, I will get the specific weather from the openweather api and display it. It was tricky because I wanted to put my searchfield component inside the form, and I did not realize that props cannot be sent up only down.  Once I figured that out, I switched the way I was handling the input and voila!